src/requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.5
django-annoying==0.10.6
django-crispy-forms==1.12.0
django-qr-code==2.2.0
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.0
pytz==2021.1
pyzbar==0.1.8
qrcode==7.1
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
